Separated comment and reaction notification tests in different files
